add parsable label to colls
